Move lib to separate project
