+++ examples/some-verifier/Cargo.toml
@@ -20,3 +20,4 @@ futures = "0.3.28"
 axum-macros = "0.3.8"
 reqwest = { version = "0.11.18", features = ["json"] }
 rust-tdlib = "0.4.3"
+some-verifier-lib = { version = "0.1", path = "../some-verifier-lib"}
